Refactor toolbar placement into navigation views
