+++ Bestuff/Sources/Stuff/Views/StuffListView.swift
@@ -22,17 +22,19 @@ struct StuffListView: View {
     @Binding private var selection: Stuff?
     @Binding private var searchText: String
 
-    @State private var sort: StuffSort = .occurredDateDescending
-    @State private var isRecapPresented = false
-    @State private var isPlanPresented = false
-    @State private var isSettingsPresented = false
-    @State private var isDebugPresented = false
+    @Binding private var sort: StuffSort
     @State private var editingStuff: Stuff?
 
-    init(stuffs: [Stuff]? = nil, selection: Binding<Stuff?>, searchText: Binding<String>) {
+    init(
+        stuffs: [Stuff]? = nil,
+        selection: Binding<Stuff?>,
+        searchText: Binding<String>,
+        sort: Binding<StuffSort>
+    ) {
         overrideStuffs = stuffs
         _selection = selection
         _searchText = searchText
+        _sort = sort
     }
 
     private var stuffs: [Stuff] {
@@ -66,68 +68,6 @@ struct StuffListView: View {
             .onDelete(perform: delete)
         }
         .navigationTitle(Text("Best Stuff"))
-        .toolbar {
-            ToolbarItem(placement: .primaryAction) {
-                AddStuffButton()
-            }
-            ToolbarItem(placement: .primaryAction) {
-                Menu {
-                    Picker("Sort", selection: $sort) {
-                        ForEach(StuffSort.allCases) { option in
-                            Text(option.title).tag(option)
-                        }
-                    }
-                } label: {
-                    Label("Sort", systemImage: "arrow.up.arrow.down")
-                }
-            }
-
-            ToolbarItem(placement: .secondaryAction) {
-                Button("Recap", systemImage: "calendar") {
-                    Logger(#file).info("Recap button tapped")
-                    isRecapPresented = true
-                }
-                .buttonStyle(.bordered)
-            }
-            ToolbarItem(placement: .secondaryAction) {
-                Button("Plan", systemImage: "lightbulb") {
-                    Logger(#file).info("Plan button tapped")
-                    isPlanPresented = true
-                }
-                .buttonStyle(.bordered)
-            }
-            ToolbarItem(placement: .secondaryAction) {
-                Button("Settings", systemImage: "gearshape") {
-                    Logger(#file).info("Settings button tapped")
-                    isSettingsPresented = true
-                }
-            }
-            #if DEBUG
-            ToolbarItem(placement: .secondaryAction) {
-                Button("Debug", systemImage: "ladybug") {
-                    Logger(#file).info("Debug button tapped")
-                    isDebugPresented = true
-                }
-            }
-            #endif
-        }
-        .sheet(isPresented: $isRecapPresented) {
-            RecapNavigationView()
-        }
-        .sheet(isPresented: $isPlanPresented) {
-            PlanNavigationView()
-        }
-        .sheet(isPresented: $isSettingsPresented) {
-            SettingsListView()
-        }
-        .sheet(isPresented: $isDebugPresented) {
-            DebugListView()
-                .toolbar {
-                    ToolbarItem(placement: .cancellationAction) {
-                        CloseButton()
-                    }
-                }
-        }
         .sheet(item: $editingStuff) { stuff in
             StuffFormView(stuff: stuff)
         }
@@ -169,7 +109,8 @@ struct StuffListView: View {
     NavigationStack {
         StuffListView(
             selection: .constant(nil),
-            searchText: .constant("")
+            searchText: .constant(""),
+            sort: .constant(.occurredDateDescending)
         )
     }
 }
